pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ies-lab"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pyyaml",
    "pandas",
    "matplotlib",
    "seaborn",
    "duckduckgo-search",
]

[project.optional-dependencies]
notebook = ["jupyter", "ipykernel"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
